+++ truffle/src/com.oracle.truffle.api.interop.java.test/src/com/oracle/truffle/api/interop/java/test/JavaInteropTest.java
@@ -189,7 +189,7 @@ public class JavaInteropTest {
         TruffleObject truffleMap = JavaInterop.asTruffleObject(map);
         TruffleObject ret = sendKeys(truffleMap);
         List<Object> list = JavaInterop.asJavaObject(List.class, ret);
-        assertThat(list, CoreMatchers.hasItems("one", "null", "three"));
+        assertThat(list, CoreMatchers.not(CoreMatchers.anyOf(CoreMatchers.hasItem("one"), CoreMatchers.hasItem("null"), CoreMatchers.hasItem("three"))));
     }
 
     @Test